+++ src/Hyper.php
@@ -56,7 +56,7 @@ class Hyper extends Plugin
     // =========================================================================
 
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '1.4.0';
+    public string $schemaVersion = '1.5.0';
 
 
     // Traits